+++ cli/cli/CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - `beam deploy` commands use solution level building instead of per-project [#3952](https://github.com/beamable/BeamableProduct/issues/3952)
-- `beam project open` command can create a `.slnf` file to show a subset of projects based on the Unity project perspective. 
+- `beam project open` command can create a `.slnf` file to show a subset of projects based on the Unity project perspective.
+
+### Fixed
+- `beam project open` works with Visual Studio
 
 ### Added
 - `beam deploy release` shows CID/PID information as part of release confirmation [#3954](https://github.com/beamable/BeamableProduct/issues/3954)